rag-service/requirements.txt
# rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# ChromaDB та векторні операції
chromadb==0.4.18
sentence-transformers==2.2.2

# Обробка документів
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
pandas==2.1.4
openpyxl==3.1.2

# HTTP запити
requests==2.31.0
httpx==0.25.2

# Логування та моніторинг
structlog==23.2.0

# Утіліти
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2